requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.30.6

# HTTP Client
httpx==0.27.2

# HTML Parsing
selectolax==0.3.17
beautifulsoup4==4.12.3
lxml==5.1.0

# Templates
jinja2==3.1.4

# Configuration
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# Scheduling
APScheduler==3.10.4

# Date/Time
pytz==2024.1
python-dateutil==2.9.0

# AWS (optional)
aioboto3==13.0.1

# Text Processing
unidecode==1.3.8
